add tumor purity logic

+++ src/gen_contaminated_bam.py
@@ -139,12 +139,79 @@ def gen_contaminated_bam(args):
         if args.cal_output_bam_coverage:
             get_coverage_from_bam(args, normal_output_bam, False, os.path.join(args.output_dir, 'cov'))
 
+        print("[INFO] Finishing merging, output file: {}".format(normal_output_bam))
+
+    #add normal to tumor
+    if tumor_purity is not None:
+        contam_coverage = tumor_bam_coverage * (1 - tumor_purity)
+        rest_tumor_coverage = tumor_bam_coverage - contam_coverage
+
+        tumor_subsample_pro = "%.3f" % (rest_tumor_coverage / float(tumor_bam_coverage))
+        normal_subsample_pro = "%.3f" % (contam_coverage / float(normal_bam_coverage))
+
+        print("[INFO] Normal/Tumor subsample proportion: {}/{}".format(normal_subsample_pro,
+                                                               tumor_subsample_pro))
+
+        tumor_subsample_bam = os.path.join(args.output_dir, 'tmp', 'tumor_rest.bam')
+        normal_subsample_bam = os.path.join(args.output_dir, 'tmp', 'normal_subsample.bam')
+
+        contig_option = "" if ctg_name is None else ctg_name
+
+        #tumor subsample cmd
+        t_s_cmd = "{} view -@{} -bh -s {} -o {} {} {}".format(samtools_execute_command,
+                                                          samtools_threads,
+                                                          tumor_subsample_pro,
+                                                          tumor_subsample_bam,
+                                                          tumor_bam_fn,
+                                                          contig_option,
+                                                          )
+        #normal subsample cmd
+        n_s_cmd = "{} view -@{} -bh -s {} -o {} {} {}".format(samtools_execute_command,
+                                                          samtools_threads,
+                                                          normal_subsample_pro,
+                                                          normal_subsample_bam,
+                                                          normal_bam_fn,
+                                                          contig_option)
+
+        n_index_cmd = '{} index -@{} {}'.format(samtools_execute_command, samtools_threads, tumor_subsample_bam)
+        t_index_cmd = '{} index -@{} {}'.format(samtools_execute_command, samtools_threads, normal_subsample_bam)
+
+
+        print('[INFO] Will run the following commands:')
+        print(t_s_cmd)
+        print(n_s_cmd)
+        print(n_index_cmd)
+        print(t_index_cmd)
+
+        if not args.dry_run:
+            subprocess.run(t_s_cmd, shell=True)
+            subprocess.run(n_s_cmd, shell=True)
+            subprocess.run(n_index_cmd, shell=True)
+            subprocess.run(t_index_cmd, shell=True)
+
+        tumor_output_bam = os.path.join(output_dir, "tumor_contaminated_{}.bam".format(tumor_purity))
+
+        print("[INFO] Merging normal BAM into tumor BAM as contamination...")
+        merge_cmd = "{} merge -f -@{} {} {} {}".format(samtools_execute_command, samtools_threads, tumor_output_bam,
+                                            tumor_subsample_bam, normal_subsample_bam)
+        index_cmd = '{} index -@{} {}'.format(samtools_execute_command, samtools_threads, tumor_output_bam)
+
+        print('[INFO] Will run the following commands:')
+        print(merge_cmd)
+        print(index_cmd)
+        if not args.dry_run:
+            subprocess_run(merge_cmd, shell=True)
+            subprocess.run(index_cmd, shell=True)
+
+        if args.cal_output_bam_coverage:
+            get_coverage_from_bam(args, tumor_output_bam, True, os.path.join(args.output_path, 'cov'))
+
         if args.remove_intermediate_dir:
             tmp_file_path = os.path.join(args.output_dir, 'tmp')
             if os.path.exist(tmp_file_path):
                 rc = subprocess.run("rm -rf {}".format(tmp_file_path), shell=True)
 
-        print("[INFO] Finishing merging, output file: {}".format(normal_output_bam))
+        print("[INFO] Finishing merging, output file: {}".format(tumor_output_bam))
 
 
 def main():
